+++ lib/scope.dart
@@ -4,7 +4,6 @@
  * Written by Brett Sutton <bsutton@onepub.dev>, Jan 2022
  */
 
-
 export 'src/exceptions.dart'
     show
         MissingDependencyException,